# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test IBID API",
        "contact": {
            "name": "Budi Haryono",
            "email": "budiharyono4@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/resreq-login": {
            "post": {
                "tags": [
                    "Authentication resreq.in (Soal no. 6)"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLoginRes",
                "requestBody": {
                    "description": "Email and Password Credential",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "eve.holt@reqres.in"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "cityslicka"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "QpwL5tke4Pnpja7X4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Credentials not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "user not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/resreq-register": {
            "post": {
                "tags": [
                    "Authentication resreq.in (Soal no. 6)"
                ],
                "summary": "Register",
                "description": "Register using email, and password",
                "operationId": "authRegisterRes",
                "requestBody": {
                    "description": "Email and Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "eve.holt@reqres.in"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pistol"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "QpwL5tke4Pnpja7X4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Only defined users succeed registration"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}